requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
opencv-python>=4.5.0
pillow>=8.0.0
matplotlib>=3.4.0
pytest>=6.2.0
pytest-cov>=2.12.0
black>=21.5b2
flake8>=3.9.0
jupyter>=1.0.0
tqdm>=4.62.0
scikit-learn>=0.24.0
pandas>=1.3.0 